// online-classroom-backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RecurrenceType = {
  NONE: 'NONE',
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
  MONTHLY_BY_DATE: 'MONTHLY_BY_DATE',
  CUSTOM: 'CUSTOM'
} as const

export type RecurrenceType = (typeof RecurrenceType)[keyof typeof RecurrenceType]


export const Day = {
  MONDAY: 'MONDAY',
  TUESDAY: 'TUESDAY',
  WEDNESDAY: 'WEDNESDAY',
  THURSDAY: 'THURSDAY',
  FRIDAY: 'FRIDAY'
} as const

export type Day = (typeof Day)[keyof typeof Day]


export const UserRole = {
  teacher: 'teacher',
  student: 'student',
  admin: 'admin'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const StudentViewPermission = {
  NO_VIEW: 'NO_VIEW',
  SCORE_ONLY: 'SCORE_ONLY',
  SCORE_AND_RESULT: 'SCORE_AND_RESULT'
} as const

export type StudentViewPermission = (typeof StudentViewPermission)[keyof typeof StudentViewPermission]


export const GradingMethod = {
  FIRST_ATTEMPT: 'FIRST_ATTEMPT',
  LATEST_ATTEMPT: 'LATEST_ATTEMPT',
  HIGHEST_ATTEMPT: 'HIGHEST_ATTEMPT'
} as const

export type GradingMethod = (typeof GradingMethod)[keyof typeof GradingMethod]


export const RequestStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]


export const NotificationType = {
  POST_LIKE: 'POST_LIKE',
  POST_COMMENT: 'POST_COMMENT',
  CLASS_APPROVAL: 'CLASS_APPROVAL',
  NEW_HOMEWORK: 'NEW_HOMEWORK',
  NEW_POST: 'NEW_POST',
  SUBMISSION_GRADED: 'SUBMISSION_GRADED',
  STUDENT_JOINED_CLASS: 'STUDENT_JOINED_CLASS'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const ClassGroupRole = {
  LEADER: 'LEADER',
  MEMBER: 'MEMBER'
} as const

export type ClassGroupRole = (typeof ClassGroupRole)[keyof typeof ClassGroupRole]